requirements.txt
asgiref==3.5.2
Django==4.1.4
django-environ==0.9.0
djangorestframework==3.14.0
environ==1.0
Markdown==3.4.1
numpy==1.24.0
psycopg==3.1.7
psycopg2-binary==2.9.5
pytz==2022.7
sqlparse==0.4.3
typing_extensions==4.4.0
